+++ rosebot/modules/images.py
@@ -1,6 +1,7 @@
 import os
 import random
 import requests
+import re
 
 from pyrogram import Filters, Message
 from rosebot.helpers import ReplyCheck
@@ -38,26 +39,28 @@ def drumpf_quote():
 
 @BOT.on_message(
     Filters.regex(
-        "(?i)(post|get|send) (rose|indigo girl|blue girl|love|roze|crush|🌹) gif"
+        "(?i)(post|get|send) (rose|indigo girl|blue girl|love|roze|crush|🌹) (gif|gifs)"
     )
 )
 def post_rose_gif(bot: BOT, message: Message):
-    image = get_random_image(ROSE_GIF_DIR)
-    BOT.send_animation(
-        chat_id=message.chat.id, animation=image, disable_notification=True, reply_to_message_id=ReplyCheck(message)
-    )
-    if message.from_user.is_self:
-        message.delete()
+    if re.match("(?i)(post|get|send) (rose|indigo girl|blue girl|love|roze|crush|🌹) (gif|gifs)", message.text):
+        image = get_random_image(ROSE_GIF_DIR)
+        BOT.send_animation(
+            chat_id=message.chat.id, animation=image, disable_notification=True, reply_to_message_id=ReplyCheck(message)
+        )
+        if message.from_user.is_self:
+            message.delete()
 
 
 @BOT.on_message(
     Filters.regex("(?i)(post|get|send) (rose|indigo girl|blue girl|love|roze|crush|🌹)")
 )
 def post_rose(bot: BOT, message: Message):
-    image = get_random_image(ROSE_DIR)
-    BOT.send_photo(chat_id=message.chat.id, photo=image, disable_notification=True, reply_to_message_id=ReplyCheck(message))
-    if message.from_user.is_self:
-        message.delete()
+    if re.match("(?i)(post|get|send) (rose|indigo girl|blue girl|love|roze|crush|🌹)", message.text):
+        image = get_random_image(ROSE_DIR)
+        BOT.send_photo(chat_id=message.chat.id, photo=image, disable_notification=True, reply_to_message_id=ReplyCheck(message))
+        if message.from_user.is_self:
+            message.delete()
 
 
 @BOT.on_message(Filters.regex("(?i)(brexit|farage|nigel)"))
@@ -68,22 +71,25 @@ def post_brexit(bot: BOT, message: Message):
 
 @BOT.on_message(Filters.regex("(?i)(post|get|send) (drumpf|trump|orange man)"))
 def post_trump(bot: BOT, message: Message):
-    image = get_random_image(TRUMP_DIR)
-    BOT.send_photo(
-        chat_id=message.chat.id,
-        photo=image,
-        disable_notification=True,
-        caption=drumpf_quote(),
-    )
-    if message.from_user.is_self:
-        message.delete()
+    if re.match("(?i)(post|get|send) (drumpf|trump|orange man)", message.text):
+        image = get_random_image(TRUMP_DIR)
+        BOT.send_photo(
+            chat_id=message.chat.id,
+            photo=image,
+            disable_notification=True,
+            caption=drumpf_quote(),
+        )
+        if message.from_user.is_self:
+            message.delete()
 
 
-@BOT.on_message(Filters.regex("(?i)(post|get|send) (tay|taytay|taylor|tswift|swift)"))
+@BOT.on_message(Filters.regex("(?i)(post|get|send) (tay|taytay|taylor|tswift|swift|trap)"))
 def post_swift(bot: BOT, message: Message):
-    image = get_random_image(SWIFT_DIR)
-    BOT.send_photo(chat_id=message.chat.id, photo=image, disable_notification=True)
-
+    if re.match("(?i)(post|get|send) (tay|taytay|taylor|tswift|swift|trap)", message.text):
+        image = get_random_image(SWIFT_DIR)
+        BOT.send_photo(chat_id=message.chat.id, photo=image, disable_notification=True)
+        if message.from_user.is_self:
+            message.delete()
 
 @BOT.on_message(
     Filters.regex(
@@ -91,73 +97,78 @@ def post_swift(bot: BOT, message: Message):
     )
 )
 def post_watson(bot: BOT, message: Message):
-    image = get_random_image(WATSON_DIR)
-    BOT.send_photo(chat_id=message.chat.id, photo=image, disable_notification=True)
-    if message.from_user.is_self:
-        message.delete()
+    if re.match("(?i)(post|get|send) (whore|slut|watson|emma|hermione|granger|harry potter girl)", message.text):
+        image = get_random_image(WATSON_DIR)
+        BOT.send_photo(chat_id=message.chat.id, photo=image, disable_notification=True)
+        if message.from_user.is_self:
+            message.delete()
 
 
 @BOT.on_message(
     Filters.regex("(?i)(post|get|send) (sminem|smnem|smn|big ear (kid|guy)|chromosome)")
 )
 def post_sminem(bot: BOT, message: Message):
-    image = get_random_image(SMINEM_DIR)
-    BOT.send_photo(chat_id=message.chat.id, photo=image, disable_notification=True)
-    if message.from_user.is_self:
-        message.delete()
+    if re.match("(?i)(post|get|send) (sminem|smnem|smn|big ear (kid|guy)|chromosome)", message.text):
+        image = get_random_image(SMINEM_DIR)
+        BOT.send_photo(chat_id=message.chat.id, photo=image, disable_notification=True)
+        if message.from_user.is_self:
+            message.delete()
 
 
 @BOT.on_message(Filters.regex("(?i)(post|get|send) (merchant|jew|rabbi)"))
 def post_merchant(bot: BOT, message: Message):
-    image = get_random_image(MERCHANT_DIR)
-    if ".gif" in os.path.splitext(image):
-        BOT.send_animation(
-            chat_id=message.chat.id,
-            animation=image,
-            disable_notification=True,
-            reply_to_message_id=ReplyCheck(message),
-        )
-    else:
-        BOT.send_photo(
-            chat_id=message.chat.id,
-            photo=image,
-            disable_notification=True,
-            reply_to_message_id=ReplyCheck(message),
-        )
-    if message.from_user.is_self:
-        message.delete()
+    if re.match("(?i)(post|get|send) (merchant|jew|rabbi)", message.text):
+        image = get_random_image(MERCHANT_DIR)
+        if ".gif" in os.path.splitext(image):
+            BOT.send_animation(
+                chat_id=message.chat.id,
+                animation=image,
+                disable_notification=True,
+                reply_to_message_id=ReplyCheck(message),
+            )
+        else:
+            BOT.send_photo(
+                chat_id=message.chat.id,
+                photo=image,
+                disable_notification=True,
+                reply_to_message_id=ReplyCheck(message),
+            )
+        if message.from_user.is_self:
+            message.delete()
 
 
 @BOT.on_message(Filters.regex("(?i)(post|install|get|send) (gentoo|linux)"))
 def post_gentoo(bot: BOT, message: Message):
-    image = get_random_image(GENTOO_DIR)
-    if ".gif" in os.path.splitext(image):
-        BOT.send_animation(
-            chat_id=message.chat.id,
-            animation=image,
-            disable_notification=True,
-            reply_to_message_id=ReplyCheck(message),
-        )
-    else:
-        BOT.send_photo(
-            chat_id=message.chat.id,
-            photo=image,
-            disable_notification=True,
-            reply_to_message_id=ReplyCheck(message),
-        )
+    if re.match("(?i)(post|install|get|send) (gentoo|linux)", message.text):
+        image = get_random_image(GENTOO_DIR)
+        if ".gif" in os.path.splitext(image):
+            BOT.send_animation(
+                chat_id=message.chat.id,
+                animation=image,
+                disable_notification=True,
+                reply_to_message_id=ReplyCheck(message),
+            )
+        else:
+            BOT.send_photo(
+                chat_id=message.chat.id,
+                photo=image,
+                disable_notification=True,
+                reply_to_message_id=ReplyCheck(message),
+            )
 
 
 @BOT.on_message(Filters.regex("(?i)(post|get|send) (redpill|meme|maymay)"))
 def post_redpill(bot: BOT, message: Message):
-    image = get_random_image(REDPILL_DIR)
-    if ".gif" in os.path.splitext(image):
-        BOT.send_animation(
-            chat_id=message.chat.id, animation=image, disable_notification=True
-        )
-    else:
-        BOT.send_photo(chat_id=message.chat.id, photo=image, disable_notification=True)
-    if message.from_user.is_self:
-        message.delete()
+    if re.match("(?i)(post|get|send) (redpill|meme|maymay)", message.text):
+        image = get_random_image(REDPILL_DIR)
+        if ".gif" in os.path.splitext(image):
+            BOT.send_animation(
+                chat_id=message.chat.id, animation=image, disable_notification=True
+            )
+        else:
+            BOT.send_photo(chat_id=message.chat.id, photo=image, disable_notification=True)
+        if message.from_user.is_self:
+            message.delete()
 
 
 @BOT.on_message(
@@ -166,12 +177,13 @@ def post_redpill(bot: BOT, message: Message):
     )
 )
 def post_911(bot: BOT, message: Message):
-    image = get_random_image(NINE_ELEVEN_DIR)
-    if ".gif" in os.path.splitext(image):
-        BOT.send_animation(
-            chat_id=message.chat.id, animation=image, disable_notification=True
-        )
-    else:
-        BOT.send_photo(chat_id=message.chat.id, photo=image, disable_notification=True)
-    if message.from_user.is_self:
-        message.delete()
+    if re.match("(?i)((post|get|send) (911|bush|twin towers|wtc)|bush did (911|9/11))", message.text):
+        image = get_random_image(NINE_ELEVEN_DIR)
+        if ".gif" in os.path.splitext(image):
+            BOT.send_animation(
+                chat_id=message.chat.id, animation=image, disable_notification=True
+            )
+        else:
+            BOT.send_photo(chat_id=message.chat.id, photo=image, disable_notification=True)
+        if message.from_user.is_self:
+            message.delete()
